fix: resolve 10 bugs across estimator, discovery, bandit, service, and frontend
- estimation.py: scalar-to-array shape mismatch in DR fold fallback (np.full)
- estimation.py: replace print() with logger.warning for propensity clip message
- discovery.py: initialise w_new/h_new to current values (not None) so outer
  loop survives when rho >= rho_max on first inner iteration
- discovery.py: guard against NaN/Inf in L-BFGS-B result before acyclicity check
- bandits.py: return int(np.argmax(p)) instead of np.int64 for type consistency
- causal_service.py: widen GRF exception catch from ImportError to Exception
  so RuntimeError from R does not crash service initialisation
- causal_service.py: log a warning instead of silently dropping out-of-bounds
  DAG edges
- mimic_pipeline.py: safe mode()[0] access guards against all-NaN column
- frontend/app.py: guard empty magnitude list before [0] index (IndexError)
- frontend/app.py: format comorbidities with :02d for fixed-width patient ID
- tests/conftest.py: coerce binomial T to int
- test_dr_local.py: coerce Stata treat column to int

+++ tests/conftest.py
@@ -23,7 +23,7 @@ def synthetic_data():
     
     logit_p = -2.0 + 0.05*(age-65) - 0.02*(bp-120) + 0.5*comorb
     p_t = np.clip(1.0 / (1.0 + np.exp(-logit_p)), 0.01, 0.99)
-    T = rng.binomial(1, p_t)
+    T = rng.binomial(1, p_t).astype(int)
     
     true_cate = 5.0 + 0.2*(age-65) - 1.5*comorb
     Y0 = 50.0 - 0.3*(age-65) + 0.1*(bp-120) - 2.0*comorb
